Moved main and settings fully into modules
